Revert "Add arguments to include/exclude sections from the generated Markdown file"
