+++ .github/workflows/ci.yaml
@@ -31,7 +31,7 @@ jobs:
     - name: Lint with golangci-lint
       uses: golangci/golangci-lint-action@v6
       with:
-        version: v1.61
+        version: v1.54
 
   test:
     name: test-${{ matrix.os }}
