+++ .gitignore
@@ -16,4 +16,3 @@
 dist/
 .DS_Store
 .vscode/
-bicep-docs-benchmark/
