+++ .pre-commit-config.yaml
@@ -10,11 +10,3 @@ repos:
     rev: v0.38.0
     hooks:
       - id: markdownlint
-  - repo: local
-    hooks:
-      - id: golint
-        name: golint
-        entry: task lint
-        language: golang
-        pass_filenames: false
-        types: [go]
