+++ go.mod
@@ -1,16 +1,14 @@
 module github.com/christosgalano/bicep-docs
 
-go 1.23
+go 1.21.6
 
 require (
 	github.com/google/uuid v1.6.0
-	github.com/json-iterator/go v1.1.12
 	github.com/spf13/cobra v1.8.1
+	golang.org/x/sync v0.7.0
 )
 
 require (
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/modern-go/concurrent v0.0.0-20180228061459-e0a39a4cb421 // indirect
-	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
 )
