--- internal/cli/testdata/README.md
@@ -1,7 +1,0 @@
-# test
-
-## Parameters
-
-| Name | Type | Description | Default |
-| --- | --- | --- | --- |
-| test_parameter | string | This is a test parameter. | "test" |
