+++ internal/markdown/markdown_test.go
@@ -10,20 +10,6 @@ import (
 	"github.com/christosgalano/bicep-docs/internal/types"
 )
 
-var (
-	defaultSections = []types.Section{
-		types.DescriptionSection,
-		types.UsageSection,
-		types.ModulesSection,
-		types.ResourcesSection,
-		types.ParametersSection,
-		types.UserDefinedDataTypesSection,
-		types.UserDefinedFunctionsSection,
-		types.VariablesSection,
-		types.OutputsSection,
-	}
-)
-
 func TestCreateFile(t *testing.T) {
 	templateName := "test"
 	templateDescription := "This is a test template."
@@ -289,7 +275,7 @@ func TestCreateFile(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			// Call CreateFile with the filename in the temporary directory
 			filename := filepath.Join(tempDir, tt.args.filename)
-			if err := CreateFile(filename, tt.args.template, false, defaultSections); (err != nil) != tt.wantErr {
+			if err := CreateFile(filename, tt.args.template, false); (err != nil) != tt.wantErr {
 				t.Errorf("CreateFile() error = %v, wantErr %v", err, tt.wantErr)
 			}
 
@@ -335,6 +321,7 @@ func Test_checkFileExists(t *testing.T) {
 		},
 	}
 	for _, tt := range tests {
+		tt := tt // Capture range variable
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 			got, err := checkFileExists(tt.args.filename)
