+++ internal/template/parse_test.go
@@ -151,7 +151,7 @@ func TestParseTemplates(t *testing.T) {
 		wantErr bool
 	}{
 		{
-			name: "basic_template",
+			name: "basic template",
 			args: args{
 				bicepFile: "testdata/basic.bicep",
 				armFile:   "testdata/basic.json",
@@ -160,7 +160,7 @@ func TestParseTemplates(t *testing.T) {
 			wantErr: false,
 		},
 		{
-			name: "extended_template",
+			name: "extended template",
 			args: args{
 				bicepFile: "testdata/extended.bicep",
 				armFile:   "testdata/extended.json",
@@ -169,7 +169,7 @@ func TestParseTemplates(t *testing.T) {
 			wantErr: false,
 		},
 		{
-			name: "non_existent_template",
+			name: "non-existent",
 			args: args{
 				bicepFile: "testdata/non-existent.bicep",
 				armFile:   "testdata/non-existent.json",
@@ -179,6 +179,7 @@ func TestParseTemplates(t *testing.T) {
 		},
 	}
 	for _, tt := range tests {
+		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 			got, err := ParseTemplates(tt.args.bicepFile, tt.args.armFile)
@@ -239,7 +240,7 @@ func Test_parseBicepTemplate(t *testing.T) {
 		wantErr       bool
 	}{
 		{
-			name: "basic",
+			name: "testdata/basic.bicep",
 			args: args{
 				bicepFile: "testdata/basic.bicep",
 			},
@@ -276,7 +277,7 @@ func Test_parseBicepTemplate(t *testing.T) {
 			wantErr:       false,
 		},
 		{
-			name: "non_existent",
+			name: "non-existent",
 			args: args{
 				bicepFile: "testdata/non-existent.bicep",
 			},
@@ -287,6 +288,7 @@ func Test_parseBicepTemplate(t *testing.T) {
 		},
 	}
 	for _, tt := range tests {
+		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 			modules, resources, variables, err := parseBicepTemplate(tt.args.bicepFile)
@@ -318,7 +320,7 @@ func Test_parseDescription(t *testing.T) {
 		want *string
 	}{
 		{
-			name: "inline_description",
+			name: "inline description",
 			args: args{
 				line:    "@description('This is a description')",
 				scanner: nil,
@@ -326,7 +328,7 @@ func Test_parseDescription(t *testing.T) {
 			want: func() *string { s := "This is a description"; return &s }(),
 		},
 		{
-			name: "multiline_description",
+			name: "multiline description",
 			args: args{
 				line:    "@sys.description('''This is a multiline ",
 				scanner: bufio.NewScanner(strings.NewReader("\ndescription\n.''' )")),
@@ -334,7 +336,7 @@ func Test_parseDescription(t *testing.T) {
 			want: func() *string { s := "This is a multiline description."; return &s }(),
 		},
 		{
-			name: "no_description",
+			name: "no description",
 			args: args{
 				line:    "This is not a description",
 				scanner: nil,
@@ -343,6 +345,7 @@ func Test_parseDescription(t *testing.T) {
 		},
 	}
 	for _, tt := range tests {
+		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 			if got := parseDescription(tt.args.line, tt.args.scanner); !reflect.DeepEqual(got, tt.want) {
@@ -363,7 +366,7 @@ func Test_parseModule(t *testing.T) {
 		want *types.Module
 	}{
 		{
-			name: "registry_source",
+			name: "registry source",
 			args: args{
 				line:        "module test 'br:exampleregistry.azurecr.io/bicep/modules/storage:v1'",
 				description: "This is a module",
@@ -374,7 +377,7 @@ func Test_parseModule(t *testing.T) {
 			},
 		},
 		{
-			name: "local_source",
+			name: "local source",
 			args: args{
 				line:        "module test './modules/test_module/main.bicep'",
 				description: "This is a module",
@@ -385,7 +388,7 @@ func Test_parseModule(t *testing.T) {
 			},
 		},
 		{
-			name: "invalid_module",
+			name: "invalid module",
 			args: args{
 				line:        "invalid line",
 				description: "This is not a module",
@@ -394,6 +397,7 @@ func Test_parseModule(t *testing.T) {
 		},
 	}
 	for _, tt := range tests {
+		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 			if got := parseModule(tt.args.line); !reflect.DeepEqual(got, tt.want) {
@@ -414,7 +418,7 @@ func Test_parseResource(t *testing.T) {
 		want *types.Resource
 	}{
 		{
-			name: "storage_account",
+			name: "storage account",
 			args: args{
 				line: "resource test 'Microsoft.Storage/storageAccounts@2023-01-01'",
 			},
@@ -424,7 +428,7 @@ func Test_parseResource(t *testing.T) {
 			},
 		},
 		{
-			name: "invalid_resource",
+			name: "invalid resource",
 			args: args{
 				line:        "invalid line",
 				description: "This is not a resource",
@@ -433,6 +437,7 @@ func Test_parseResource(t *testing.T) {
 		},
 	}
 	for _, tt := range tests {
+		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 			if got := parseResource(tt.args.line); !reflect.DeepEqual(got, tt.want) {
@@ -453,7 +458,7 @@ func Test_skipComment(t *testing.T) {
 		args args
 	}{
 		{
-			name: "single_line_comment",
+			name: "single line comment",
 			args: args{
 				line:     "// This is a single line comment",
 				scanner:  nil,
@@ -469,7 +474,7 @@ func Test_skipComment(t *testing.T) {
 			},
 		},
 		{
-			name: "normal_line",
+			name: "normal line",
 			args: args{
 				line:     "This is a normal line",
 				scanner:  nil,
@@ -478,6 +483,7 @@ func Test_skipComment(t *testing.T) {
 		},
 	}
 	for _, tt := range tests {
+		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 			skipped, err := skipComment(tt.args.line, tt.args.scanner)
