+++ internal/types/unmarshal.go
@@ -1,36 +1,18 @@
 package types
 
 import (
+	"encoding/json"
 	"fmt"
 	"strings"
-
-	jsoniter "github.com/json-iterator/go"
 )
-
-var json = jsoniter.ConfigCompatibleWithStandardLibrary
-
-func unmarshalTypeOrRef(data []byte) (string, error) {
-	var result struct {
-		Type string `json:"type"`
-		Ref  string `json:"$ref"`
-	}
-	if err := json.Unmarshal(data, &result); err != nil {
-		return "", err
-	}
-	if result.Type != "" {
-		return result.Type, nil
-	}
-	if result.Ref != "" {
-		return result.Ref, nil
-	}
-	return "", fmt.Errorf("neither type nor $ref field found")
-}
 
 // UnmarshalJSON unmarshals a JSON object into a Parameter.
 // The type field can be either a type or a $ref.
 func (p *Parameter) UnmarshalJSON(data []byte) error {
 	type Alias Parameter
 	aux := &struct {
+		Type json.RawMessage `json:"type"`
+		Ref  json.RawMessage `json:"$ref"`
 		*Alias
 	}{
 		Alias: (*Alias)(p),
@@ -39,13 +21,19 @@ func (p *Parameter) UnmarshalJSON(data []byte) error {
 		return err
 	}
 
-	tr, err := unmarshalTypeOrRef(data)
-	if err != nil {
-		return err
+	var ref string
+	if err := json.Unmarshal(aux.Ref, &ref); err == nil {
+		p.Type = ref
+		return nil
 	}
-	p.Type = tr
 
-	return nil
+	var typ string
+	if err := json.Unmarshal(aux.Type, &typ); err == nil {
+		p.Type = typ
+		return nil
+	}
+
+	return fmt.Errorf("type field could not be unmarshalled")
 }
 
 // UnmarshalJSON unmarshals a JSON object into an Output.
@@ -53,6 +41,8 @@ func (p *Parameter) UnmarshalJSON(data []byte) error {
 func (o *Output) UnmarshalJSON(data []byte) error {
 	type Alias Output
 	aux := &struct {
+		Type json.RawMessage `json:"type"`
+		Ref  json.RawMessage `json:"$ref"`
 		*Alias
 	}{
 		Alias: (*Alias)(o),
@@ -61,13 +51,19 @@ func (o *Output) UnmarshalJSON(data []byte) error {
 		return err
 	}
 
-	tr, err := unmarshalTypeOrRef(data)
-	if err != nil {
-		return err
+	var ref string
+	if err := json.Unmarshal(aux.Ref, &ref); err == nil {
+		o.Type = ref
+		return nil
 	}
-	o.Type = tr
 
-	return nil
+	var typ string
+	if err := json.Unmarshal(aux.Type, &typ); err == nil {
+		o.Type = typ
+		return nil
+	}
+
+	return fmt.Errorf("type field could not be unmarshalled")
 }
 
 // UnmarshalJSON unmarshals a JSON object into a UserDefinedDataType.
@@ -75,6 +71,8 @@ func (o *Output) UnmarshalJSON(data []byte) error {
 func (u *UserDefinedDataType) UnmarshalJSON(data []byte) error {
 	type Alias UserDefinedDataType
 	aux := &struct {
+		Type json.RawMessage `json:"type"`
+		Ref  json.RawMessage `json:"$ref"`
 		*Alias
 	}{
 		Alias: (*Alias)(u),
@@ -83,13 +81,19 @@ func (u *UserDefinedDataType) UnmarshalJSON(data []byte) error {
 		return err
 	}
 
-	tr, err := unmarshalTypeOrRef(data)
-	if err != nil {
-		return err
+	var ref string
+	if err := json.Unmarshal(aux.Ref, &ref); err == nil {
+		u.Type = ref
+		return nil
 	}
-	u.Type = tr
 
-	return nil
+	var typ string
+	if err := json.Unmarshal(aux.Type, &typ); err == nil {
+		u.Type = typ
+		return nil
+	}
+
+	return fmt.Errorf("type field could not be unmarshalled")
 }
 
 // UnmarshalJSON unmarshals a JSON object into a Template.
